implemented control flow
Implemented control‑flow stack unwinding with block result propagation while preserving operand stacks for void function exits, added global type validation, expanded arch macros, and updated branch tests to match the new stack semantics.

Details

fa_runtime.c adds a preserve_stack control‑frame flag, skips unwinding for implicit void function blocks, and uses the flag in end/branch handling for proper stack unwinding.
fa_ops.c validates global.get/global.set against declared global types before pushing/storing.
fa_arch.h now exposes override‑friendly pointer size, endianness, and CPU family detection macros.
fa_wasm.c fixes size_read initialization in global initializer parsing.
main.c switches branch tests to result‑typed blocks, carries br_table results explicitly, and adds a forward declaration to avoid implicit function warnings.
README.md and AI_REFERENCE.md reflect the updated runtime behavior and gaps

+++ src/fa_types.h
@@ -15,7 +15,7 @@ typedef short           i16;
 typedef uint8_t         u8;
 typedef int8_t          i8;
 
-typedef uint64_t fa_ptr;
+typedef uintptr_t fa_ptr;
 typedef void* ptr; // just the current system pointer
 
 // Essential callbacks
